requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
pydantic>=2.0.0
spacy>=3.7.2
pdfplumber>=0.11.0
wordfreq>=3.0.0
numpy>=1.24.0
pytest>=7.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.0
python-dotenv>=0.21.0
